pyproject.toml
[tool.poetry]
name = "side-by-side-bom"
version = "0.1.0"
description = ""
authors = ["jonnyjackson26 <jrsjackson26@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-docx = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
